Fix memory leaks and type safety in notification service
- Replace BehaviorSubject anti-pattern with of() operator to prevent memory leaks
- Add NotificationSettings interface with proper type definitions
- Add error handling to requestDesktopPermission() Promise
- Improve code documentation and remove console.log statements
- Update component to use proper NotificationSettings type instead of any

These changes address critical code review findings from PR #24.
